+++ src/routes.jsx
@@ -21,7 +21,7 @@ import JoinTeam from "./containers/JoinTeam";
 import Home from "./containers/Home";
 import Settings from "./containers/Settings";
 import UserEdit from "./containers/UserEdit";
-import TexterFaqs from "./components/TexterFaqs";
+import TexterFaqs from "./components/TexterFrequentlyAskedQuestions";
 import FAQs from "./lib/faqs";
 
 export default function makeRoutes(requireAuth = () => {}) {
